small updates

+++ python/vtool/maya_lib/util.py
@@ -669,6 +669,21 @@ class MayaVariable(vtool.util.Variable):
         return cmds.objExists(self._get_node_and_variable())
 
     #--- set
+    def set_name(self, name):
+        
+        var_name = self._get_node_and_variable()
+        
+        print var_name
+        
+        if cmds.objExists(var_name):
+            cmds.renameAttr(var_name, name)
+            
+        super(MayaVariable, self).set_name(name)
+        
+        var_name = self._get_node_and_variable()
+        print var_name, cmds.getAttr(var_name)
+        
+    
     def set_value(self, value):
         super(MayaVariable, self).set_value(value)
         self._set_value()
@@ -921,6 +936,9 @@ class Attributes(object):
         
         node_and_attribute = '%s.%s' % (self.node, attribute_name)
         
+        if not cmds.objExists(node_and_attribute):
+            return
+        
         var_type = cmds.getAttr(node_and_attribute, type = True)
         
         var = self._get_variable_instance(attribute_name, var_type)
@@ -1029,12 +1047,18 @@ class Attributes(object):
     
     def rename_variable(self, old_name, new_name):
         
+        if not self.node:
+            return
+        
         var = self.get_variable(old_name)
+        
+        if not var:
+            vtool.util.warning('Could not rename attribute, %s.%s.' % (self.node, old_name))
+            return
         
         var.set_name(new_name)
         
-        #here it needs to rename the attribute in maya as well!!!!!!!!!!!
-        #!!!!!!!!!!!!!
+        self._store_attributes()
         
         self.delete_all()
         self.create_all()
